requirements.txt
locustio
ipython
pyzmq
